backend/requirements.txt
pytest
ruff
mypy
black
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
sqlalchemy[asyncio]  # For SQLAlchemy with asyncio support
asyncpg              # PostgreSQL driver for SQLAlchemy (for Cloud SQL)
aiosqlite            # SQLite driver for SQLAlchemy (for local dev fallback)
passlib[bcrypt]      # For password hashing
python-jose[cryptography] # For JWT tokens (can also use PyJWT)
